Hydrocarbons
This is mostly a personal code for learning as a beginner. Its purpose is translating the technical name of hydrocarbons and simplifying it. I may share an update for this in some time since my goal is learning how i can make the machine draw the formula of the hydrocarbon using its name.